Made the plpgsql sum work
But it is even slower than native sum. Unless it is implemented in
C/C++.

Either find a new way to do this, which is not likely, or just create
wrappers for MADlib's matrix multiplication.

There are multiple important bugs fixes in this branch.

Also it would be better to let as.db.data.frame just create temp tables
if no table_name is given.