Two changes:
Care is taken to avoid coefficients diverging in order to achieve
fitted probabilities of 0. When a probability is within 1e-5 of 1,
set it to 1, and set the weight to be 1e-5. 0 side is treated similarly.

Also use nargs in replacement methods instead of sys.calls to avoid the
interfere of environment