Start using the new version policy, Version 0.1.10.13
All versions on CRAN have 3 numbers: xx.xx.xx

After updating the CRAN version, master begins to accumulate the 4th
number and the version number on master would be xx.xx.xx.xx, where the
first 3 numbers are the same to the current CRAN version. The 4th number
starts from 1 and stops increasing right before uploading the next CRAN
version. It would increase whenever a new commit that changes the
functionality or the user doc is pushed into master.

Right before the next uploading to CRAN, the third number would be
updated and the version would be again xx.xx.xx. The package is ready
for uploading to CRAN again.

In this way, the user, who is using the master version, can easily
distinguish it from other versions by using
'packageVersion("PivotalR")' command in R console.