Testing: Update to new testthat requirements
Closes #48

The old way of building test cases were to run the PivotalR:::test
function. This function was built upon the testthat module and included
some extra functionality. It needed to be updated regularly with changes
in testthat. To avoid such changes, this commit removes the function and
changes the test process to directly use testthat.

The downside of this change is that the test user needs to create global
environment variables instead of providing the port and database name
through system input. Also we lose the capability to test the doc
examples.