Update add GP6 and GP7 support, fix warnings and CRAN reqs
GP6 and GP7 have slightly different catalog tables compared to GP5.
This commit makes the changes for compatibility of all three GP
versions.

The suppress warnings function was using "panic" as default which
is incorrect. This commit fixes it for consistent warning messages.

CRAN requirements have changed since the last release and we fix
a number of minor issues to comply with the rules.